Refactored throw away email service to be encapsulated inside TestAccount for easy refactoring later if we want to change services.
