+++ src/main/groovy/com/stormpath/tck/authentication/CookieIT.groovy
@@ -26,6 +26,7 @@ import static com.stormpath.tck.util.CookieUtils.isCookieDeleted
 import static com.stormpath.tck.util.FrameworkConstants.LoginRoute
 import static com.stormpath.tck.util.FrameworkConstants.MeRoute
 import static com.stormpath.tck.util.Matchers.urlStartsWithPath
+import static com.stormpath.tck.util.TestAccount.Mode.WITHOUT_DISPOSABLE_EMAIL
 import static org.hamcrest.Matchers.is
 import static org.hamcrest.Matchers.not
 import static org.testng.Assert.assertEquals
@@ -97,7 +98,7 @@ class CookieIT extends AbstractIT {
      */
     @Test(groups=["v100", "json"])
     void cookieExpirationMatchesTokenTtl() throws Exception {
-        def account = new TestAccount()
+        def account = new TestAccount(WITHOUT_DISPOSABLE_EMAIL)
         account.registerOnServer()
         deleteOnClassTeardown(account.href)
