+++ src/main/groovy/com/stormpath/tck/forgot/ForgotPasswordIT.groovy
@@ -29,6 +29,7 @@ import static com.jayway.restassured.RestAssured.given
 import static com.jayway.restassured.RestAssured.put
 import static com.stormpath.tck.util.FrameworkConstants.ForgotRoute
 import static com.stormpath.tck.util.Matchers.urlMatchesPath
+import static com.stormpath.tck.util.TestAccount.Mode.WITHOUT_DISPOSABLE_EMAIL
 import static org.hamcrest.MatcherAssert.assertThat
 import static org.hamcrest.Matchers.allOf
 import static org.hamcrest.Matchers.isEmptyOrNullString
@@ -37,7 +38,7 @@ import static org.testng.Assert.assertEquals
 
 class ForgotPasswordIT extends AbstractIT {
 
-    private TestAccount account = new TestAccount()
+    private TestAccount account = new TestAccount(WITHOUT_DISPOSABLE_EMAIL)
     private static final invalidEmail = "foo+notarealemail@bar.baz"
 
     @BeforeClass
